# lib/promrelabel: speedup label match by `__name__`